refactor: My -> EditProfileModal 리팩토링 완료
